+++ zarrs/src/array/codec/array_to_bytes/sharding.rs
@@ -68,6 +68,8 @@ mod sharding_options;
 mod sharding_partial_decoder_async;
 mod sharding_partial_decoder_sync;
 mod sharding_partial_encoder;
+#[cfg(feature = "async")]
+mod sharding_partial_encoder_async;
 
 #[cfg(feature = "async")]
 pub use sharding_partial_decoder_async::AsyncShardingPartialDecoder;